+++ CHANGELOG.md
@@ -33,12 +33,18 @@
   - Added annotated types in compatibility with the [annotated-types](https://github.com/annotated-types/annotated-types) package
   - Added `Run.external_service` (recently added by server, but purpose is unclear)
   - Notes about possible deprecation removed and type hints updated for some fields as usages, albeit rare, have been found
+  - `VehiclePosition.direction` can be `None`
+  - Added new possible value for `Disruption.disruption_type`
 - Fixed bugs in `ptv_timetable.TimetableAPI` and `ptv_timetable.asyncapi.AsyncTimetableAPI`:
   - `geopath_date` value in `get_route()` was not converted to string before being sent to the API
   - A `None` argument check for the `date` parameter was missing in `get_run()`
   - Some URL parameter names in `get_fare_estimate()` were incorrect
   - `get_fare_estimate()` can return `None`
   - Corrected overload signatures in `search()`
+- Fixed bugs in `ptv_timetable.types`:
+  - `RunInterchange` was absent from `__all__`
+  - `Run.interchange` `dict` values weren't converted to `RunInterchage` instances
+- Added some unit tests
 - Updated documentation
 
 ## 0.4.1 (2025-03-17)